javascript commit
When focused is changed the calcart function is ran. When onchange occurs the calCart function is run. provides validation test for selection list. Calculates the cost of the registration and saves data in session storage. when the fields are being used by the user, the information will be entered when the user switches the field. Writes data values from session storage in to the registration summary form. the users information is put on the page when entered. copied the write session Values function over to the reg2 javascript file.